requirements.txt
Flask==2.3.2
email-validator==1.3.1
python-dotenv==1.0.0